move more actions into platform specific. split out avatar stuff (#12686)
